requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.2.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
